last chance

+++ test.php
@@ -1,23 +1,3 @@
-<html>
-<head>
-<meta charset="utf-8">
-<title>Test page</title>
-<style>
-    body {
-        width: 75%;
-        margin: 0 auto;
-        font-family: Tahoma, Verdana, Arial, sans-serif;
-    }
-</style>
-</head>
-<body>
-<table border="1px" border-collapse="collapse" width="100%">
-  <tr>
-    <td align="center"><a href="admin.php"><span>Файл admin.php</span></a></td>
-	<td align="center"><a href="list.php"><span>Файл list.php</span></a></td>
-	<td align="center"><span>Файл test.php</span></td>
-  </tr>
- </table>
 <?php
 if (!empty($_GET)){
 	if (empty($_GET['username'])){
@@ -31,22 +11,28 @@ if (!empty($_GET)){
 		exit;
 	}
 	else{
-		$json = file_get_contents($_GET['test'].'.json');
-		$data = json_decode($json, true);
-		$username = $_GET['username'];
-		
-		echo '<p>Выполните тест:</p>'; 
-		echo '<form method="post" action="test.php">';
-		foreach($data as $element){
-			echo '<p><b>'.$element['Question'].'</b></p>';
-			foreach($element['Answers'] as $answer){
-				echo '<p><input type="radio" name="'.$element['id'].'" value="a1">'.$answer['1'].'<br>';
-				echo '<p><input type="radio" name="'.$element['id'].'" value="a2">'.$answer['2'].'<br>';
-				echo '<p><input type="radio" name="'.$element['id'].'" value="a3">'.$answer['3'].'<br>';
-			}
+		if(!file_exists($_GET['test'].'.json')){
+			header("HTTP/1.0 404 Not Found");
 		}
-		echo '<p><input type="submit" value="Отправить ответы"></p>';
-		echo '</form>';
+		else{
+			$json = file_get_contents($_GET['test'].'.json');
+			$data = json_decode($json, true);
+			
+			echo '<p>Выполните тест:</p>'; 
+			echo '<form method="post" action="test.php">';
+			foreach($data as $element){
+				echo '<p><b>'.$element['Question'].'</b></p>';
+				foreach($element['Answers'] as $answer){
+					echo '<p><input type="radio" name="'.$element['id'].'" value="a1">'.$answer['1'].'<br>';
+					echo '<p><input type="radio" name="'.$element['id'].'" value="a2">'.$answer['2'].'<br>';
+					echo '<p><input type="radio" name="'.$element['id'].'" value="a3">'.$answer['3'].'<br>';
+				}
+			}
+			echo '<input type="hidden" name="username" id="username" value="'.$_GET['username'].'" />';
+			echo '<p><input type="submit" value="Отправить ответы"></p>';
+			echo '</form>';
+		}
+
 	}
 }
 
@@ -58,13 +44,14 @@ else if(!empty($_POST)){
 		if ($answer == "a2"){
 			$n += 1;
 		}
-	}
-	header("Content-type: image/png");
-	$string = $username.', поздравляем!'.'<br>'.'Вы набрали '.$n.'баллов!';
+	}	
+	
+	$string = $_POST["username"].', ваше количество баллов: '.$n;
 	$im     = imagecreatefrompng("certificate.png");
 	$orange = imagecolorallocate($im, 220, 210, 60);
-	$px     = (imagesx($im) - 7.5 * strlen($string)) / 2;
-	imagestring($im, 3, $px, 190, $string, $orange);
+	$font = 'hermanomayor.ttf';
+	imagettftext ($im, 16 , 0 , 130, 190 , $orange , $font , $string);
+	header("Content-type: image/png; charset=utf-8");
 	imagepng($im);
 	imagedestroy($im);
 }
@@ -74,5 +61,3 @@ else if(empty($_GET)){
 	exit;
 }
  ?>
-</body>
-</html>
